+++ NEWS.md
@@ -1,4 +1,5 @@
 ## 5.9.0 - Unreleased
+* [MODINVOICE-543](https://folio-org.atlassian.net/browse/MODINVOICE-543) - EDIFACT invoice lines load out of order
 
 ## 5.8.0 - Released (Quesnelia R1 2024)
 The focus of this release was to fix bugs and make improvement in transaction call and error codes