+++ NEWS.md
@@ -1,5 +1,13 @@
 ## 5.9.0 - Unreleased
 
+## 5.8.3 - Released (Quesnelia R1 2024)
+
+This release focused on fixing security vulnerabilities
+[Full Changelog](https://github.com/folio-org/mod-invoice/compare/v5.8.2...v5.8.3)
+
+### Bug Fixes
+* [MODINVOICE-558](https://folio-org.atlassian.net/browse/MODINVOICE-558) - Fix Vert.x, Netty, Apache SSHD/SFTP - CVE-2024-41909
+
 ## 5.8.2 - Released (Quesnelia R1 2024)
 This release focused on adding ability to search by location and holding in POL  
 [Full Changelog](https://github.com/folio-org/mod-invoice/compare/v5.8.1...v5.8.2)